// FrontEnd/src/app/models/event.ts
export interface Event {
    eventId?: number; // Optional if auto-generated by backend
    eventTitle: string;
    eventDescription?: string;
    eventDate?: string; // Use string for simplicity, or Date if you handle parsing
    category?: Category;
    eventLocation?: string;
    places?: number; // New property for number of places
    // Add other properties as per your backend Event entity
}

export enum Category {
    SHOW="SHOW", PARTY="PARTY", EXCURSION="EXCURSION", FORUM="FORUM", SPORT="SPORT"

}